新增新區域 新增增益效果 新增技能 新增物品 新增敵人
調整第一層區域設定
調整戰鬥邏輯
調整增益處理邏輯
調整鍛造邏輯

+++ Assets/Scripts/Client/Item/ItemEnemy.cs
@@ -33,7 +33,7 @@ public class ItemEnemy : MonoBehaviour
         iconSelected.SetActive(isOn);
     }
 
-    public void GetDamage(int damage)
+    public void GetDamage(decimal damage)
     {
         Info.CharacterData.CurrentHP -= damage;
         if (Info.CharacterData.CurrentHP < 0)
